feat: scale break speed with a bot's own tool tier
Upstream advanced one fixed crack stage every two ticks, so every block took
twenty ticks whatever the bot held and whatever it was breaking. That made a
tool tier decoration. Break progress is now the held tool's destroy speed
against the block, so netherite takes 14 ticks and wood takes 60.

STAGE_COST is defined as iron's progress in one run — speed 6.0 over a
two-tick period, so 12 — which makes iron still take exactly twenty ticks by
construction rather than by coincidence. A GameTest asserts that number
exactly, ticking only the scheduler: ticking the registry would run the
agent, the agent moves bots, and the break task cancels itself the moment a
bot stops aiming at its block. Block hardness stays ignored, as upstream
ignored it.

The tools slot floors at wood. An empty hand scores 1.0 against everything,
which would be 120 ticks a block as the result of skipping an argument;
`none` still parses there because it is the create chain's filler word. The
clamp lives in Bot.setToolTier so no caller can route around it, with
EquipmentTier.asToolTier as its single definition.

Two consequences. AgentState.mining widens from Byte to Short — diamond's
step is 16, so progress reaches 128 before it is checked and a byte tops out
at 127. And the UNBREAKABLE refusal moves above the destroy branch: upstream
checked it afterwards, which was safe only because an unbreakable block never
advanced a stage and so never reached nine, and a netherite bot would now
overshoot straight past it into bedrock.

+++ src/main/java/net/nuggetmc/tplus/agent/AgentState.java
@@ -78,8 +78,15 @@ public final class AgentState {
      */
     public final Map<BlockRef, Short> crackList = new HashMap<>();
 
-    /** Mining task id to progress stage, 0 through 9. */
-    public final Map<Integer, Byte> mining = new HashMap<>();
+    /**
+     * Mining task id to accumulated break progress, 0 to {@code Mining.BREAK_COST}.
+     *
+     * <p>Was the crack stage, 0 through 9, when every block took a fixed twenty ticks. It is now
+     * progress, because a stage is no longer a fixed number of ticks — see
+     * {@code Mining.blockBreakEffect}. {@code Short} rather than {@code Byte} because a fast tool
+     * can push progress past 127 in the run that breaks the block.
+     */
+    public final Map<Integer, Short> mining = new HashMap<>();
 
     /** Bots that took fall damage recently and must not be shoved downward again. */
     public final Set<Bot> fallDamageCooldown = new HashSet<>();
